feat: big chunks done
