+++ packages/amplify-category-auth/src/provider-utils/awscloudformation/index.js
@@ -3,202 +3,32 @@ const open = require('open');
 const path = require('path');
 const fs = require('fs-extra');
 const _ = require('lodash');
-const uuid = require('uuid');
-const { existsSync } = require('fs');
-const { copySync } = require('fs-extra');
 const { getAuthResourceName } = require('../../utils/getAuthResourceName');
-const { ServiceName: FunctionServiceName } = require('amplify-category-function');
-
-// Todo: move these to supported service.json
-
-const ENV_SPECIFIC_PARAMS = [
-  'facebookAppId',
-  'facebookAppIdUserPool',
-  'facebookAppSecretUserPool',
-  'googleClientId',
-  'googleIos',
-  'googleAndroid',
-  'googleAppIdUserPool',
-  'googleAppSecretUserPool',
-  'amazonAppId',
-  'loginwithamazonAppIdUserPool',
-  'loginwithamazonAppSecretUserPool',
-  'hostedUIProviderCreds',
-];
-
-const safeDefaults = [
-  'allowUnauthenticatedIdentities',
-  'thirdPartyAuth',
-  'authProviders',
-  'smsAuthenticationMessage',
-  'emailVerificationSubject',
-  'emailVerificationMessage',
-  'smsVerificationMessage',
-  'passwordPolicyMinLength',
-  'passwordPolicyCharacters',
-  'userpoolClientRefreshTokenValidity',
-];
-
-const protectedValues = [
-  'resourceName',
-  'userPoolName',
-  'identityPoolName',
-  'usernameAttributes',
-  'autoVerifiedAttributes',
-  'requiredAttributes',
-];
-
-const privateKeys = [
-  'facebookAppIdUserPool',
-  'facebookAuthorizeScopes',
-  'facebookAppSecretUserPool',
-  'googleAppIdUserPool',
-  'googleAuthorizeScopes',
-  'googleAppSecretUserPool',
-  'loginwithamazonAppIdUserPool',
-  'loginwithamazonAuthorizeScopes',
-  'loginwithamazonAppSecretUserPool',
-  'CallbackURLs',
-  'LogoutURLs',
-  'AllowedOAuthFlows',
-  'AllowedOAuthScopes',
-  'EditURLS',
-  'newCallbackURLs',
-  'addCallbackOnUpdate',
-  'updateFlow',
-  'newCallbackURLs',
-  'selectedParties',
-  'newLogoutURLs',
-  'editLogoutURLs',
-  'addLogoutOnUpdate',
-  'additionalQuestions',
-];
-
-const resourcesRoot = path.normalize(path.join(__dirname, '../../../resources'));
-const adminAuthAssetRoot = path.join(resourcesRoot, 'adminAuth');
-const cfnTemplateRoot = path.join(resourcesRoot, 'cloudformation-templates');
+const { verificationBucketName } = require('./utils/verification-bucket-name');
+const {
+  copyCfnTemplate,
+  removeDeprecatedProps,
+  createUserPoolGroups,
+  addAdminAuth,
+  saveResourceParameters,
+  lambdaTriggers,
+  copyS3Assets,
+} = require('./utils/synthesize-resources');
+const { protectedValues, safeDefaults, ENV_SPECIFIC_PARAMS, privateKeys } = require('./constants');
+const { getAddAuthHandler } = require('./handlers/get-add-auth-handler');
 
 function serviceQuestions(context, defaultValuesFilename, stringMapFilename, serviceWalkthroughFilename, serviceMetadata) {
   const serviceWalkthroughSrc = `${__dirname}/service-walkthroughs/${serviceWalkthroughFilename}`;
   const { serviceWalkthrough } = require(serviceWalkthroughSrc);
-
   return serviceWalkthrough(context, defaultValuesFilename, stringMapFilename, serviceMetadata);
 }
 
-async function copyCfnTemplate(context, category, options, cfnFilename) {
-  const { amplify } = context;
-  const targetDir = amplify.pathManager.getBackendDirPath();
-
-  const copyJobs = [
-    {
-      dir: cfnTemplateRoot,
-      template: cfnFilename,
-      target: `${targetDir}/${category}/${options.resourceName}/${options.resourceName}-cloudformation-template.yml`,
-      paramsFile: `${targetDir}/${category}/${options.resourceName}/parameters.json`,
-    },
-  ];
-
-  const privateParams = Object.assign({}, options);
-  privateKeys.forEach(p => delete privateParams[p]);
-
-  return await context.amplify.copyBatch(context, copyJobs, privateParams, true);
-}
-
-function saveResourceParameters(context, providerName, category, resource, params, envSpecificParams = []) {
-  const provider = context.amplify.getPluginInstance(context, providerName);
-  let privateParams = Object.assign({}, params);
-  privateKeys.forEach(p => delete privateParams[p]);
-  privateParams = removeDeprecatedProps(privateParams);
-  provider.saveResourceParameters(context, category, resource, privateParams, envSpecificParams);
-}
-
-async function addResource(context, category, service) {
-  let props = {};
+async function addResource(context, service) {
   const serviceMetadata = require('../supported-services').supportedServices[service];
-  const { cfnFilename, defaultValuesFilename, stringMapFilename, serviceWalkthroughFilename, provider } = serviceMetadata;
-  let projectName = context.amplify.getProjectConfig().projectName.toLowerCase();
-  const disallowedChars = /[^A-Za-z0-9]+/g;
-  projectName = projectName.replace(disallowedChars, '');
-
-  return serviceQuestions(context, defaultValuesFilename, stringMapFilename, serviceWalkthroughFilename, serviceMetadata)
-    .then(async (result) => {
-      const defaultValuesSrc = `${__dirname}/assets/${defaultValuesFilename}`;
-      const { functionMap, generalDefaults, roles } = require(defaultValuesSrc);
-
-      /* if user has used the default configuration,
-       * we populate base choices like authSelections and resourceName for them */
-      if (['default', 'defaultSocial'].includes(result.useDefault)) {
-        result = Object.assign(generalDefaults(projectName), result);
-      }
-
-      await verificationBucketName(result);
-
-      /* merge actual answers object into props object,
-       * ensuring that manual entries override defaults */
-      props = Object.assign(functionMap[result.authSelections](result.resourceName), result, roles);
-
-      await lambdaTriggers(props, context, null);
-
-      await createUserPoolGroups(context, props.resourceName, result.userPoolGroupList);
-
-      await addAdminAuth(context, props.resourceName, 'add', result.adminQueryGroup);
-
-      await copyCfnTemplate(context, category, props, cfnFilename);
-
-      saveResourceParameters(context, provider, category, result.resourceName, props, ENV_SPECIFIC_PARAMS);
-    })
-    .then(async () => {
-      await copyS3Assets(context, props);
-      return props.resourceName;
-    });
-}
-
-async function createUserPoolGroups(context, resourceName, userPoolGroupList) {
-  if (userPoolGroupList && userPoolGroupList.length > 0) {
-    const userPoolGroupPrecedenceList = [];
-
-    for (let i = 0; i < userPoolGroupList.length; i += 1) {
-      userPoolGroupPrecedenceList.push({
-        groupName: userPoolGroupList[i],
-        precedence: i + 1,
-      });
-    }
-
-    const userPoolGroupFile = path.join(
-      context.amplify.pathManager.getBackendDirPath(),
-      'auth',
-      'userPoolGroups',
-      'user-pool-group-precedence.json',
-    );
-
-    const userPoolGroupParams = path.join(context.amplify.pathManager.getBackendDirPath(), 'auth', 'userPoolGroups', 'parameters.json');
-
-    /* eslint-disable */
-    const groupParams = {
-      AuthRoleArn: {
-        'Fn::GetAtt': ['AuthRole', 'Arn'],
-      },
-      UnauthRoleArn: {
-        'Fn::GetAtt': ['UnauthRole', 'Arn'],
-      },
-    };
-    /* eslint-enable */
-
-    fs.outputFileSync(userPoolGroupParams, JSON.stringify(groupParams, null, 4));
-    fs.outputFileSync(userPoolGroupFile, JSON.stringify(userPoolGroupPrecedenceList, null, 4));
-
-    context.amplify.updateamplifyMetaAfterResourceAdd('auth', 'userPoolGroups', {
-      service: 'Cognito-UserPool-Groups',
-      providerPlugin: 'awscloudformation',
-      dependsOn: [
-        {
-          category: 'auth',
-          resourceName,
-          attributes: ['UserPoolId', 'AppClientIDWeb', 'AppClientID', 'IdentityPoolId'],
-        },
-      ],
-    });
-  }
+  const { defaultValuesFilename, stringMapFilename, serviceWalkthroughFilename } = serviceMetadata;
+  return getAddAuthHandler(context)(
+    await serviceQuestions(context, defaultValuesFilename, stringMapFilename, serviceWalkthroughFilename, serviceMetadata),
+  );
 }
 
 // may be able to consolidate this into just createUserPoolGroups
@@ -603,284 +433,11 @@ function getPermissionPolicies(context, service, resourceName, crudOptions) {
   return getIAMPolicies(resourceName, crudOptions);
 }
 
-async function lambdaTriggers(coreAnswers: { [x: string]: any;  }, context: { amplify: { getTriggerPermissions: (arg0: any, arg1: any, arg2: string, arg3: any) => any; pathManager: { getBackendDirPath: () => any; }; deleteAllTriggers: (arg0: any, arg1: any, arg2: any, arg3: any) => any; dependsOnBlock: (arg0: any, arg1: string[], arg2: string) => any; }; }, previouslySaved: {} | null) {
-  const { handleTriggers } = require('./utils/trigger-flow-auth-helper');
-  let triggerKeyValues = {};
-
-  if (coreAnswers.triggers) {
-    triggerKeyValues = await handleTriggers(context, coreAnswers, previouslySaved);
-    coreAnswers.triggers = triggerKeyValues ? JSON.stringify(triggerKeyValues) : '{}';
-
-    if (triggerKeyValues) {
-      coreAnswers.parentStack = { Ref: 'AWS::StackId' };
-    }
-
-    // determine permissions needed for each trigger module
-    coreAnswers.permissions = await context.amplify.getTriggerPermissions(context, coreAnswers.triggers, 'auth', coreAnswers.resourceName);
-  } else if (previouslySaved) {
-    const targetDir = context.amplify.pathManager.getBackendDirPath();
-    Object.keys(previouslySaved).forEach(p => {
-      delete coreAnswers[p];
-    });
-    await context.amplify.deleteAllTriggers(previouslySaved, coreAnswers.resourceName, targetDir, context);
-  }
-  // remove unused coreAnswers.triggers key
-  if (coreAnswers.triggers && coreAnswers.triggers === '[]') {
-    delete coreAnswers.triggers;
-  }
-
-  // handle dependsOn data
-  const dependsOnKeys = Object.keys(triggerKeyValues).map(i => `${coreAnswers.resourceName}${i}`);
-  coreAnswers.dependsOn = context.amplify.dependsOnBlock(context, dependsOnKeys, 'Cognito');
-}
-
-async function copyS3Assets(context, props) {
-  const targetDir = `${context.amplify.pathManager.getBackendDirPath()}/auth/${props.resourceName}/assets`;
-
-  const triggers = props.triggers ? JSON.parse(props.triggers) : null;
-  const confirmationFileNeeded = props.triggers && triggers.CustomMessage && triggers.CustomMessage.includes('verification-link');
-  if (confirmationFileNeeded) {
-    if (!existsSync(targetDir)) {
-      const source = path.join(__dirname, `../../../provider-utils/awscloudformation/triggers/CustomMessage/assets`);
-      copySync(source, `${targetDir}`);
-    }
-  }
-}
-
-async function verificationBucketName(current, previous?) {
-  if (current.triggers && current.triggers.CustomMessage && current.triggers.CustomMessage.includes('verification-link')) {
-    const name = previous ? previous.resourceName : current.resourceName;
-    current.verificationBucketName = `${name.toLowerCase()}verificationbucket`;
-  } else if (
-    previous &&
-    previous.triggers &&
-    previous.triggers.CustomMessage &&
-    previous.triggers.CustomMessage.includes('verification-link') &&
-    previous.verificationBucketName &&
-    (!current.triggers || !current.triggers.CustomMessage || !current.triggers.CustomMessage.includes('verification-link'))
-  ) {
-    delete previous.updatingAuth.verificationBucketName;
-  }
-}
-
-function removeDeprecatedProps(props) {
-  if (props.authRoleName) {
-    delete props.authRoleName;
-  }
-  if (props.unauthRoleName) {
-    delete props.unauthRoleName;
-  }
-  if (props.userpoolClientName) {
-    delete props.userpoolClientName;
-  }
-  if (props.roleName) {
-    delete props.roleName;
-  }
-  if (props.policyName) {
-    delete props.policyName;
-  }
-  if (props.mfaLambdaLogPolicy) {
-    delete props.mfaLambdaLogPolicy;
-  }
-  if (props.mfaPassRolePolicy) {
-    delete props.mfaPassRolePolicy;
-  }
-  if (props.mfaLambdaIAMPolicy) {
-    delete props.mfaLambdaIAMPolicy;
-  }
-  if (props.userpoolClientLogPolicy) {
-    delete props.userpoolClientLogPolicy;
-  }
-  if (props.userpoolClientLambdaPolicy) {
-    delete props.userpoolClientLambdaPolicy;
-  }
-  if (props.lambdaLogPolicy) {
-    delete props.lambdaLogPolicy;
-  }
-  if (props.openIdRolePolicy) {
-    delete props.openIdRolePolicy;
-  }
-  if (props.openIdLambdaIAMPolicy) {
-    delete props.openIdLambdaIAMPolicy;
-  }
-  if (props.openIdLogPolicy) {
-    delete props.openIdLogPolicy;
-  }
-  if (props.mfaLambdaRole) {
-    delete props.mfaLambdaRole;
-  }
-  if (props.openIdLambdaRoleName) {
-    delete props.openIdLambdaRoleName;
-  }
-
-  const keys = Object.keys(props);
-  const deprecatedTriggerParams = [
-    'CreateAuthChallenge',
-    'CustomMessage',
-    'DefineAuthChallenge',
-    'PostAuthentication',
-    'PostConfirmation',
-    'PreAuthentication',
-    'PreSignup',
-    'VerifyAuthChallengeResponse',
-  ];
-  keys.forEach(el => {
-    if (deprecatedTriggerParams.includes(el)) {
-      delete props[el];
-    }
-  });
-  return props;
-}
-
-async function addAdminAuth(context, authResourceName, operation, adminGroup, functionName) {
-  if (adminGroup) {
-    if (!functionName) {
-      const [shortId] = uuid().split('-');
-      functionName = `AdminQueries${shortId}`;
-    }
-    await createAdminAuthFunction(context, authResourceName, functionName, adminGroup, operation);
-    await createAdminAuthAPI(context, authResourceName, functionName, operation);
-  }
-}
-
-async function createAdminAuthFunction(context, authResourceName, functionName, adminGroup, operation) {
-  const targetDir = context.amplify.pathManager.getBackendDirPath();
-  const pluginDir = __dirname;
-  let lambdaGroupVar = adminGroup;
-
-  const dependsOn = [];
-
-  dependsOn.push({
-    category: 'auth',
-    resourceName: authResourceName,
-    attributes: ['UserPoolId'],
-  });
-
-  if (!lambdaGroupVar) {
-    lambdaGroupVar = 'NONE';
-  }
-
-  const functionProps = {
-    functionName: `${functionName}`,
-    roleName: `${functionName}LambdaRole`,
-    dependsOn,
-    authResourceName,
-    lambdaGroupVar,
-  };
-
-  const copyJobs = [
-    {
-      dir: adminAuthAssetRoot,
-      template: 'admin-auth-app.js',
-      target: `${targetDir}/function/${functionName}/src/app.js`,
-    },
-    {
-      dir: adminAuthAssetRoot,
-      template: 'admin-auth-cognitoActions.js',
-      target: `${targetDir}/function/${functionName}/src/cognitoActions.js`,
-    },
-    {
-      dir: adminAuthAssetRoot,
-      template: 'admin-auth-index.js',
-      target: `${targetDir}/function/${functionName}/src/index.js`,
-    },
-    {
-      dir: adminAuthAssetRoot,
-      template: 'admin-auth-package.json',
-      target: `${targetDir}/function/${functionName}/src/package.json`,
-    },
-    {
-      dir: adminAuthAssetRoot,
-      template: 'admin-queries-function-template.json.ejs',
-      target: `${targetDir}/function/${functionName}/${functionName}-cloudformation-template.json`,
-    },
-  ];
-
-  // copy over the files
-  await context.amplify.copyBatch(context, copyJobs, functionProps, true);
-
-  if (operation === 'add') {
-    // add amplify-meta and backend-config
-    const backendConfigs = {
-      service: FunctionServiceName.LambdaFunction,
-      providerPlugin: 'awscloudformation',
-      build: true,
-      dependsOn,
-    };
-
-    await context.amplify.updateamplifyMetaAfterResourceAdd('function', functionName, backendConfigs);
-    context.print.success(`Successfully added ${functionName} function locally`);
-  } else {
-    context.print.success(`Successfully updated ${functionName} function locally`);
-  }
-}
-
-async function createAdminAuthAPI(context, authResourceName, functionName, operation) {
-  const targetDir = context.amplify.pathManager.getBackendDirPath();
-  const pluginDir = __dirname;
-  const apiName = 'AdminQueries';
-  const dependsOn = [];
-
-  dependsOn.push(
-    {
-      category: 'auth',
-      resourceName: authResourceName,
-      attributes: ['UserPoolId'],
-    },
-    {
-      category: 'function',
-      resourceName: functionName,
-      attributes: ['Arn', 'Name'],
-    },
-  );
-
-  const apiProps = {
-    functionName,
-    authResourceName,
-    dependsOn,
-  };
-
-  const copyJobs = [
-    {
-      dir: adminAuthAssetRoot,
-      template: 'admin-queries-api-template.json.ejs',
-      target: `${targetDir}/api/${apiName}/admin-queries-cloudformation-template.json`,
-    },
-    {
-      dir: adminAuthAssetRoot,
-      template: 'admin-queries-api-params.json',
-      target: `${targetDir}/api/${apiName}/parameters.json`,
-    },
-  ];
-
-  // copy over the files
-  await context.amplify.copyBatch(context, copyJobs, apiProps, true);
-
-  if (operation === 'add') {
-    // Update amplify-meta and backend-config
-    const backendConfigs = {
-      service: 'API Gateway',
-      providerPlugin: 'awscloudformation',
-      dependsOn,
-    };
-
-    await context.amplify.updateamplifyMetaAfterResourceAdd('api', apiName, backendConfigs);
-    context.print.success(`Successfully added ${apiName} API locally`);
-  } else {
-    context.print.success(`Successfully updated ${apiName} API locally`);
-  }
-}
-
 module.exports = {
   addResource,
   updateResource,
   updateConfigOnEnvInit,
-  saveResourceParameters,
-  ENV_SPECIFIC_PARAMS,
-  copyCfnTemplate,
   migrate,
   console,
   getPermissionPolicies,
-  removeDeprecatedProps,
-  cfnTemplateRoot,
 };
