+++ packages/amplify-category-auth/src/provider-utils/awscloudformation/service-walkthroughs/auth-questions.js
@@ -198,6 +198,7 @@ async function serviceWalkthrough(context, defaultValuesFilename, stringMapsFile
     ...coreAnswers,
     userPoolGroupList,
     adminQueryGroup,
+    serviceName: 'Cognito',
   };
 }
 
@@ -707,4 +708,5 @@ module.exports = {
   parseOAuthCreds,
   structureoAuthMetaData,
   getIAMPolicies,
+  identityPoolProviders,
 };
