+++ packages/amplify-category-auth/src/provider-utils/awscloudformation/utils/verification-bucket-name.ts
@@ -1,0 +1,15 @@
+export const verificationBucketName = async (current: any, previous?: any) => {
+  if (current.triggers && current.triggers.CustomMessage && current.triggers.CustomMessage.includes('verification-link')) {
+    const name = previous ? previous.resourceName : current.resourceName;
+    current.verificationBucketName = `${name.toLowerCase()}verificationbucket`;
+  } else if (
+    previous &&
+    previous.triggers &&
+    previous.triggers.CustomMessage &&
+    previous.triggers.CustomMessage.includes('verification-link') &&
+    previous.verificationBucketName &&
+    (!current.triggers || !current.triggers.CustomMessage || !current.triggers.CustomMessage.includes('verification-link'))
+  ) {
+    delete previous.updatingAuth.verificationBucketName;
+  }
+};
